Move v4 utils to a dedicated module
